Introducing BookEditionCategory
From now on, we are using a relationship between Books, Editions and Categories to make sure we have a sustainable way of having that on permanent record, while enabling publishers to change each book's categories for future editions.

Adding a rake task that creates BookEditionCategory records.

Adding basic validation to books.

Also adding the `for_web` and `for_print` attributes to BookCategory as those are needed now.