Reverting preloaded cover image preloads
Preloading is not as effective as I thought. Besides, the viewport size seems to be unknown before rendering so the largest image size is preloaded instead of the appropriate one for the viewport.

But it is no biggie, as lazy loading images below the fold is more effective than preloading large static images.