Removing eager loading for cover images
So it turns out that eager loading for ActiveStorage attachments was "fixed" in Rails version 7.1.  This led to an insanely large and complex query being called for the books index view.

Some testing shows that it makes much more sense to skip the eager loading and querying the datbase for each cover image, even if it results in an n+1 query.