Fixing seeds and development state
As I managed to crash my development database and was unable to restore it from a production database dump due to conflicts, the best course of action was to rebuild the development seeding mechanism.

This also facilitates other things such as not crashing when the database is missing some information.