Eliminating `.with_attached_cover_image` statements
From version 7.1 of Rails, calling `.with_attached_cover_image` queries the database for every single image variation in the database instead of the pervious expected behaviour, which was to only query for basic information about each ActiveStorage attachment.

This has caused massive delays in processing as well as timeout errors, so we are removing the eager loading statement and accepting the fact that we'll have a couple of n+1 queries instead.